backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
openai
pydantic
requests
python-multipart
PyPDF2
python-docx
neo4j
